refactor(discussions): split the lambda into layered modules
The discussions lambda was a single 1457-line index.js — the largest in
the repo. Split it along the layers it already had internally, keeping
behavior identical (handler + close are still exported from index.js, and
the load-time takeover-safety invariant still fires on import):

- constants.js   — literals + the takeover-safety invariant assertion
- clients.js     — shared Neptune (query/close) + DDB/SSM/Lambda clients
- mappers.js     — pure helpers and Neptune->DTO mappers (no I/O)
- data-access.js — Neptune graph + DynamoDB reads/writes
- services.js    — sprint authorization, WS fan-out, secret resolution
- handlers.js    — one HTTP handler per route
- index.js       — slim router (1457 -> 94 lines)

No logic changed; this is a pure move + re-export. All 64 existing tests
pass unchanged, the esbuild bundle still builds (it follows the new ./
imports), lint and format are clean.
